fix(HIGH-24): query, mutation 파일명 변경
